Fix BFloat16 type error on v_loss
I don’t know why 7ecb62781046e2384209d8f2b484b7796b092b24 appears to have caused this though

+++ rl_algo_impls/ppo/ppo.py
@@ -351,7 +351,7 @@ class PPO(Algorithm):
                     TrainStepStats(
                         loss.item(),
                         pi_loss.item(),
-                        v_loss.detach().cpu().numpy(),
+                        v_loss.detach().float().cpu().numpy(),
                         entropy_loss.item(),
                         approx_kl,
                         clipped_frac,
